Added a broken hack for "=" return expressions.
Now, the "=" short hand can be used to return values.
It is intended to later return values from statements, and turn
statements into expressions, but for now, it is equivalent to a RETURN,
so it doesn't work as intended, but for returning from a function body,
it works just fine.